Tighten laterality test assertions and move to test_projection_schema.py per review
